Re-do structs to support UDM w/ UAP.
