+++ .travis.yml
@@ -11,5 +11,5 @@ before_install:
 install:
 - dep ensure
 script:
-- golangci-lint run --enable-all -e G402 -e dupl
+- golangci-lint run --enable-all -e G402
 - go test ./...
